pyproject.toml
[project]
name = "ai-agent-demo"
version = "1.0.0"
description = "Portfolio-grade FastAPI multi-agent orchestration demo with real tool-calling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "prometheus-client>=0.20.0",
  "opentelemetry-api>=1.27.0",
  "opentelemetry-sdk>=1.27.0",
  "opentelemetry-exporter-otlp>=1.27.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
